% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sdtm.R
\docType{data}
\name{SUPPAE}
\alias{SUPPAE}
\title{Supplemental Qualifiers for Adverse Events (SUPPAE)}
\format{
A data frame with 3 rows and 8 variables:
\describe{
\item{STUDYID}{Study identifier}
\item{RDOMAIN}{Related domain}
\item{USUBJID}{Unique subject identifier}
\item{IDVAR}{Identifying variable name}
\item{IDVARVAL}{Identifying variable value}
\item{QNAM}{Qualifier variable name}
\item{QLABEL}{Qualifier label}
\item{QVAL}{Qualifier value}
}
}
\source{
Simulated SDTM-like data
}
\usage{
SUPPAE
}
\description{
A minimal SDTM SUPPAE dataset containing supplemental
qualifiers for the AE domain.
}
\keyword{datasets}
